# api/openapi.yaml
components:
  schemas:
...
    diagnostichttp.AnswerResponse:
      additionalProperties: false
      properties:
        competency_id:
          type: string
        created_at:
          type: integer
        criterion_results:
          items:
            $ref: '#/components/schemas/diagnostichttp.CriterionResultResponse'
          type: array
          uniqueItems: false
        feedback:
          items:
            type: string
          type: array
          uniqueItems: false
        grader_max_score:
          maximum: 2
          minimum: 1
          type: integer
        grader_score:
          maximum: 2
          minimum: 0
          type: integer
        outcome_id:
          type: string
        role:
          enum:
            - main
            - basic
          type: string
        score:
          maximum: 2
          minimum: 0
          type: integer
        source_task_id:
          type: string
        task:
          $ref: '#/components/schemas/diagnostichttp.TaskResponse'
        text:
          type: string
        transcription_id:
          type: string
        variant_task_id:
          type: string
        verdict:
          enum:
            - correct
            - partial
            - incorrect
          type: string
      required:
        - competency_id
        - created_at
        - criterion_results
        - feedback
        - grader_max_score
        - grader_score
        - outcome_id
        - role
        - score
        - source_task_id
        - task
        - text
        - transcription_id
        - variant_task_id
        - verdict
      type: object
    diagnostichttp.AudioMetadataResponse:
      additionalProperties: false
      properties:
        audio_url:
          type:
            - string
            - "null"
        status:
          enum:
            - missing
            - pending
            - processing
            - ready
            - failed
            - cancelled
          type: string
        variant_task_id:
          type: string
      required:
        - audio_url
        - status
        - variant_task_id
      type: object
    diagnostichttp.CriterionResultResponse:
      additionalProperties: false
      properties:
        explanation:
          type: string
        key:
          type: string
        satisfied:
          type: boolean
      required:
        - explanation
        - key
        - satisfied
      type: object
    diagnostichttp.ProgressResponse:
      additionalProperties: false
      properties:
        completed_tasks:
          minimum: 0
          type: integer
        criterion_results:
          items:
            $ref: '#/components/schemas/diagnostichttp.CriterionResultResponse'
          type: array
          uniqueItems: false
        current:
          $ref: '#/components/schemas/diagnostichttp.TaskResponse'
        feedback:
          items:
            type: string
          type: array
          uniqueItems: false
        grader_max_score:
          maximum: 2
          minimum: 1
          type: integer
        grader_score:
          maximum: 2
          minimum: 0
          type: integer
        score:
          maximum: 2
          minimum: 0
          type: integer
        session_id:
          maxLength: 128
          minLength: 1
          type: string
        skipped_tasks:
          minimum: 0
          type: integer
        status:
          enum:
            - active
            - completed
          type: string
        text:
          minLength: 1
          type: string
        total_tasks:
          minimum: 1
          type: integer
        verdict:
          enum:
            - correct
            - partial
            - incorrect
          type: string
      required:
        - completed_tasks
        - session_id
        - skipped_tasks
        - status
        - total_tasks
      type: object
    diagnostichttp.ResultResponse:
      additionalProperties: false
      properties:
        answers:
          items:
            $ref: '#/components/schemas/diagnostichttp.AnswerResponse'
          type: array
          uniqueItems: false
        completed_tasks:
          type: integer
        diagnostic_score:
          type: integer
        included_competency_count:
          type: integer
        map_revision:
          type: integer
        maximum_score:
          type: integer
        session_id:
          type: string
        skipped_competencies:
          items:
            $ref: '#/components/schemas/diagnostichttp.SkippedCompetencyResponse'
          type: array
          uniqueItems: false
        status:
          enum:
            - completed
          type: string
        total_tasks:
          type: integer
        untested_basics:
          items:
            $ref: '#/components/schemas/diagnostichttp.TaskResponse'
          type: array
          uniqueItems: false
        variant_id:
          type: string
      required:
        - answers
        - completed_tasks
        - diagnostic_score
        - included_competency_count
        - map_revision
        - maximum_score
        - session_id
        - skipped_competencies
        - status
        - total_tasks
        - untested_basics
        - variant_id
      type: object
    diagnostichttp.SkippedCompetencyResponse:
      additionalProperties: false
      properties:
        code:
          type: string
        competency_id:
          type: string
        competency_name:
          type: string
      required:
        - code
        - competency_id
        - competency_name
      type: object
    diagnostichttp.StartRequest:
      additionalProperties: false
      properties:
        variant_id:
          maxLength: 128
          minLength: 1
          type: string
      required:
        - variant_id
      type: object
    diagnostichttp.TaskResponse:
      additionalProperties: false
      properties:
        competency_id:
          maxLength: 128
          minLength: 1
          type: string
        competency_name:
          type: string
        constituent_id:
          maxLength: 128
          minLength: 1
          type: string
        constituent_name:
          type: string
        options:
          items:
            type: string
          type: array
          uniqueItems: false
        outcome_id:
          maxLength: 128
          minLength: 1
          type: string
        outcome_name:
          type: string
        question:
          minLength: 1
          type: string
        role:
          enum:
            - main
            - basic
          type: string
        source_task_id:
          maxLength: 128
          minLength: 1
          type: string
        variant_task_id:
          maxLength: 128
          minLength: 1
          type: string
        voice_instruction:
          maxLength: 500
          minLength: 1
          type: string
      required:
        - competency_id
        - competency_name
        - constituent_id
        - constituent_name
        - options
        - outcome_id
        - outcome_name
        - question
        - role
        - source_task_id
        - variant_task_id
        - voice_instruction
      type: object
    fault.Detail:
      additionalProperties: false
      properties:
        code:
          type: string
        message:
          type: string
        path:
          type: string
      required:
        - code
        - message
      type: object
    fault.Error:
      additionalProperties: false
      properties:
        code:
          type: string
        details:
          items:
            $ref: '#/components/schemas/fault.Detail'
          type: array
          uniqueItems: false
        message:
          type: string
      required:
        - code
        - message
      type: object
...
  securitySchemes:
    accessCookie:
      in: cookie
      name: access_token
      type: apiKey
    refreshCookie:
      in: cookie
      name: refresh_token
      type: apiKey
info:
  description: 'Единый API: implemented — подключённые операции, planned — будущие операции. Variantgen собирает, сохраняет и читает варианты; грейдинг, ответы и тренировка находятся вне его scope. Источник — текущий ML CSV/XLSX. Даты — Unix seconds; cookies Secure/HttpOnly.'
  title: AI Tutor Backend API
  version: 0.2.0
openapi: 3.1.0
paths:
...
  /diagnostic-sessions:
    post:
      description: Сохраняет в памяти снимок порядка варианта. Повтор с тем же ключом возвращает ту же сессию.
      parameters:
        - description: Ключ повтора запроса
          in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/diagnostichttp.StartRequest'
              description: Вариант для прохождения
              summary: request
        description: Вариант для прохождения
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/diagnostichttp.ProgressResponse'
          description: OK
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/diagnostichttp.ProgressResponse'
          description: Created
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Unprocessable Entity
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Service Unavailable
      security:
        - accessCookie: []
      summary: Начать диагностическую сессию
      tags:
        - Диагностика
      x-implementation: implemented
  /diagnostic-sessions/{id}:
    get:
      parameters:
        - description: ID сессии
          in: path
          name: id
          required: true
          schema:
            maxLength: 128
            minLength: 1
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/diagnostichttp.ProgressResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Not Found
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Unprocessable Entity
      security:
        - accessCookie: []
      summary: Прочитать состояние диагностики
      tags:
        - Диагностика
      x-implementation: implemented
  /diagnostic-sessions/{id}/answers:
    post:
      description: Выполняет STT, затем вызывает assessment application с variant_id, variant_task_id и transcription_id. Грейдер читает доверенный снимок и роль; ошибка оценки не меняет позицию.
      parameters:
        - description: ID сессии
          in: path
          name: id
          required: true
          schema:
            maxLength: 128
            minLength: 1
            type: string
        - description: Ключ повтора запроса
          in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              additionalProperties: false
              properties:
                audio:
                  description: WAV, Ogg/Opus или WebM/Opus, до 25 МиБ
                  format: binary
                  type: string
                variant_task_id:
                  description: ID текущей позиции варианта
                  type: string
              required:
                - audio
                - variant_task_id
              type: object
        description: WAV, Ogg/Opus или WebM/Opus, до 25 МиБ
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/diagnostichttp.ProgressResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Conflict
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Request Entity Too Large
        "415":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Unsupported Media Type
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Unprocessable Entity
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Bad Gateway
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Gateway Timeout
      security:
        - accessCookie: []
      summary: Отправить голосовой ответ
      tags:
        - Диагностика
      x-implementation: implemented
  /diagnostic-sessions/{id}/current/audio:
    get:
      parameters:
        - description: ID сессии
          in: path
          name: id
          required: true
          schema:
            maxLength: 128
            minLength: 1
            type: string
        - description: ID текущего задания
          in: query
          name: variant_task_id
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/diagnostichttp.AudioMetadataResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Unprocessable Entity
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Service Unavailable
      security:
        - accessCookie: []
      summary: Прочитать состояние озвучки текущего задания
      tags:
        - Диагностика
      x-implementation: implemented
...
  /diagnostic-sessions/{id}/result:
    get:
      parameters:
        - description: ID сессии
          in: path
          name: id
          required: true
          schema:
            maxLength: 128
            minLength: 1
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/diagnostichttp.ResultResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Unprocessable Entity
      security:
        - accessCookie: []
      summary: Прочитать результат диагностики
      tags:
        - Диагностика
      x-implementation: implemented
...
  /task-audio/{id}/file:
    get:
      parameters:
        - description: ID аудиозаписи задания
          in: path
          name: id
          required: true
          schema:
            maxLength: 256
            minLength: 1
            type: string
      responses:
        "200":
          content:
            audio/wav:
              schema:
                format: binary
                type: string
          description: Сохранённая WAV-озвучка
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Unprocessable Entity
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fault.Error'
          description: Service Unavailable
      security:
        - accessCookie: []
      summary: Получить сохранённую озвучку задания
      tags:
        - Задания
      x-implementation: implemented
...
servers:
  - description: Локальный API через Caddy (Docker Compose)
    url: https://localhost:8443/api/v1
  - description: HTTPS reverse proxy
    url: /api/v1
...